==> notebooks/01_pandas_basics.ipynb <==
import pandas as pd

print("pandas successfully loaded")

import seaborn as sns

df = sns.load_dataset("iris")

df.head()

df.info()

df.shape

df.describe()

df.columns

df['species'].value_counts()

df['sepal_length'].head()

df[['sepal_length', 'petal_length']].head()

df[df['species'] == 'setosa'].head()

df[df['sepal_length'] > 6.0].head()

#new ratio col
df['sepal_ratio'] = df['sepal_length'] / df['sepal_width']

df.head()

#grouping and summarizing

df.groupby("species")[['sepal_length', 'petal_length']].mean()

#saving and reading from a csv
path = r"C:\Ibrahim\Personal\University Stuff\Machine Learning\Project\ML Irrigation Project\data\processed\sample_pandas_output.csv"

df.to_csv(path, index=False)

df2 = pd.read_csv(r"C:\Ibrahim\Personal\University Stuff\Machine Learning\Project\ML Irrigation Project\data\processed\sample_pandas_output.csv")
df2.head()
