==> notebooks/07_process_merged_data.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv(r"C:\Ibrahim\Personal\University Stuff\Machine Learning\Project\ML Irrigation Project\data\raw\final_district_dataset.csv")

df.head(10)

df.tail(10)

df.shape

#remove duplicate district-year rows
dup = df.groupby(["District", "Year"]).size().sort_values(ascending=False)

dup.head(20)

df = df.groupby(["District", "Year"], as_index=False).agg({
    "Avg_Rainfall":"mean",
    "Area_Sown":"mean",
    "Wheat_Percentage":"mean",
    "Avg_Temperature":"mean",
    "Crop_Yield":"mean",
    "Irrigation_Area":"mean"
})

df.head(15)

df.shape

#cleaning numeric cols
num_cols = ["Avg_Rainfall","Area_Sown","Wheat_Percentage","Avg_Temperature",
            "Crop_Yield","Irrigation_Area"]

for col in num_cols:
    df[col] = pd.to_numeric(df[col], errors="coerce")

#handling missing vals

df.isna().sum()

df[num_cols] = df[num_cols].fillna(df[num_cols].median())

df.isna().sum()

df.shape

df.head()

df.tail(10)

df.duplicated().sum()

df = df.drop(columns={"Area_Sown", "Wheat_Percentage"})

df.columns.tolist()

df.shape

df.to_csv(r"C:\Ibrahim\Personal\University Stuff\Machine Learning\Project\ML Irrigation Project\data\processed\final_cleaned_district_dataset.csv", index=False)
