==> notebooks/10_geospatial_feature_engineering.ipynb <==
import pandas as pd
import geopandas as gpd
from shapely.geometry import Polygon

shapefile_path = r"C:\Ibrahim\Personal\University Stuff\Machine Learning\Project\ML Irrigation Project\data\geospatial\pakistan_shapefile\gadm41_PAK_2.shp"

gdf = gpd.read_file(shapefile_path)

gdf.head()

#our clean dataset
df = pd.read_csv(
    r"C:\Ibrahim\Personal\University Stuff\Machine Learning\Project\ML Irrigation Project\data\processed\final_cleaned_district_dataset.csv"
)

df.head()

gdf[['GID_2', 'NAME_2']].head(20)

#match codes
gdf["DistrictCode"] = gdf["GID_2"].str.extract(r'\.(\d+)\.')
gdf["DistrictCode"] = "PK" + gdf["DistrictCode"]

#merge both datasets
geo_merged = gdf.merge(df, left_on="DistrictCode", right_on="District")

geo_merged.head()
